pyproject.toml
[project]
name = "argusnet"
version = "0.1.0"
description = "ArgusNet - Modular Network Diagnostics Toolkit"
authors = [
    { name = "Pedro de Faria" }
]
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12.0",
    "aiohttp>=3.9.0",
    "dnspython>=2.4.0",
    "python-whois>=0.8.0",
    "aiodns>=3.0.0",
    "requests>=2.31.0",
    "geoip2>=4.7.0",
    "entropy>=0.1.0",
    "cryptography>=41.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
argusnet = "argusnet.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
